requirements.txt
datasets>=2.19.0
openai>=1.37.0
anthropic>=0.36.0
tiktoken>=0.7.0
pydantic>=2.7.0
pandas>=2.2.2
matplotlib>=3.9.0
numpy>=1.26.4
pyyaml>=6.0.1
tqdm>=4.66.4
rich>=13.7.1
tenacity>=8.3.0
